Applies authenticator to admin views; Minor tweaks in templates
