+++ photolog/templates/login.html
@@ -1,6 +1,6 @@
 <h1>In</h1>
   <form action="" method="post" name="login">
-      {{ form.hidden_tag() }}  
+      {{ form.hidden_tag() }}
     <p>
     {{ form.username.label }}:<br>
       {{ form.username(size=25) }}<br>
