+++ photolog/templates/logout.html
@@ -1,1 +1,3 @@
 <h1>Bye Bye</h1>
+
+<a href="{{ url_for('index') }}">Home</a>
